+++ CHANGES.md
@@ -1,1 +1,4 @@
-- removed obvious references to NimPlant in Server and Client headers
+<h1>Changes for Stage 2</h1>
+
+<h2>Server</h2>
+- Added multiple artificial server names that change randomly to hide the keyword ''NimPlant Cs Server''